requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
langchain-groq
pillow
streamlit